+++ CHANGELOG.md
@@ -48,3 +48,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### [1.0.4] - yusufw2429@gmail.com
+#### Added
+- add wishlist
+- add parse token jwt to context
+- add price on product
+- tidy up routes